refactored row_set and row_range to separate files
