+++ tensorflow_io/core/kernels/bigtable/bigtable_dataset_kernel.cc
@@ -21,6 +21,7 @@ limitations under the License.
 #include "tensorflow/core/framework/op_kernel.h"
 #include "tensorflow/core/framework/resource_mgr.h"
 #include "tensorflow/core/framework/resource_op_kernel.h"
+#include "tensorflow_io/core/kernels/bigtable/bigtable_row_set.h"
 
 using ::tensorflow::DT_STRING;
 using ::tensorflow::PartialTensorShape;
@@ -298,371 +299,6 @@ class BigtableDatasetOp : public DatasetOpKernel {
 
 REGISTER_KERNEL_BUILDER(Name("BigtableDataset").Device(DEVICE_CPU),
                         BigtableDatasetOp);
-
-class BigtableRowsetResource : public ResourceBase {
- public:
-  explicit BigtableRowsetResource(cbt::RowSet const& row_set) {
-    VLOG(1) << "BigtableRowsetResource ctor";
-    row_set_ = std::move(row_set);
-  }
-
-  ~BigtableRowsetResource() { VLOG(1) << "BigtableRowsetResource dtor"; }
-
-  std::string ToString() const {
-    std::string res;
-    google::protobuf::TextFormat::PrintToString(row_set_.as_proto(), &res);
-    return res;
-  }
-
-  void AppendRow(std::string const& row_key) { row_set_.Append(row_key); }
-  void AppendRowRange(cbt::RowRange const& row_range) {
-    row_set_.Append(row_range);
-  }
-  cbt::RowSet Intersect(cbt::RowRange const& row_range) {
-    return row_set_.Intersect(row_range);
-  }
-
-  string DebugString() const override {
-    return "BigtableRowsetResource:{" + ToString() + "}";
-  }
-
- private:
-  cbt::RowSet row_set_;
-};
-
-class BigtableRowRangeResource : public ResourceBase {
- public:
-  explicit BigtableRowRangeResource(cbt::RowRange row_range)
-      : row_range_(std::move(row_range)) {
-    VLOG(1) << "BigtableRowsetResource ctor";
-  }
-
-  ~BigtableRowRangeResource() { VLOG(1) << "BigtableRowsetResource dtor"; }
-
-  std::string ToString() const {
-    std::string res;
-    google::protobuf::TextFormat::PrintToString(row_range_.as_proto(), &res);
-    return res;
-  }
-
-  cbt::RowRange& RowRange() { return row_range_; }
-
-  string DebugString() const override {
-    return "BigtableRowRangeResource:{" + ToString() + "}";
-  }
-
- private:
-  cbt::RowRange row_range_;
-};
-
-template <typename T>
-class OpKernelCreatingResource : public OpKernel {
- public:
-  explicit OpKernelCreatingResource(OpKernelConstruction* context)
-      : OpKernel(context) {}
-
-  void Compute(OpKernelContext* context) override TF_LOCKS_EXCLUDED(mu_) {
-    ResourceMgr* mgr = context->resource_manager();
-    ContainerInfo cinfo;
-    OP_REQUIRES_OK(context, cinfo.Init(mgr, def()));
-
-    T* resource;
-    OP_REQUIRES_OK(context, CreateResource(&resource));
-
-    OP_REQUIRES_OK(context,
-                   mgr->Create<T>(cinfo.container(), cinfo.name(), resource));
-
-    OP_REQUIRES_OK(context, MakeResourceHandleToOutput(
-                                context, 0, cinfo.container(), cinfo.name(),
-                                TypeIndex::Make<T>()));
-  }
-
- private:
-  virtual Status CreateResource(T** resource)
-      TF_EXCLUSIVE_LOCKS_REQUIRED(mu_) = 0;
-};
-
-class BigtableEmptyRowRangeOp
-    : public OpKernelCreatingResource<BigtableRowRangeResource> {
- public:
-  explicit BigtableEmptyRowRangeOp(OpKernelConstruction* ctx)
-      : OpKernelCreatingResource<BigtableRowRangeResource>(ctx) {
-    VLOG(1) << "BigtableEmptyRowRangeOp ctor ";
-  }
-
- private:
-  Status CreateResource(BigtableRowRangeResource** resource)
-      TF_EXCLUSIVE_LOCKS_REQUIRED(mu_) override {
-    *resource = new BigtableRowRangeResource(cbt::RowRange::Empty());
-    return Status::OK();
-  }
-
- private:
-  mutable mutex mu_;
-};
-
-REGISTER_KERNEL_BUILDER(Name("BigtableEmptyRowRange").Device(DEVICE_CPU),
-                        BigtableEmptyRowRangeOp);
-
-class BigtableRowRangeOp
-    : public OpKernelCreatingResource<BigtableRowRangeResource> {
- public:
-  explicit BigtableRowRangeOp(OpKernelConstruction* ctx)
-      : OpKernelCreatingResource<BigtableRowRangeResource>(ctx) {
-    VLOG(1) << "BigtableRowRangeOp ctor ";
-    OP_REQUIRES_OK(ctx, ctx->GetAttr("left_row_key", &left_row_key_));
-    OP_REQUIRES_OK(ctx, ctx->GetAttr("left_open", &left_open_));
-    OP_REQUIRES_OK(ctx, ctx->GetAttr("right_row_key", &right_row_key_));
-    OP_REQUIRES_OK(ctx, ctx->GetAttr("right_open", &right_open_));
-  }
-
- private:
-  Status CreateResource(BigtableRowRangeResource** resource)
-      TF_EXCLUSIVE_LOCKS_REQUIRED(mu_) override {
-    VLOG(1) << "BigtableRowRangeOp constructing row_range:"
-            << (left_open_ ? "(" : "[") << left_row_key_ << ":"
-            << right_row_key_ << (right_open_ ? ")" : "]");
-
-    // both empty - infinite
-    if (left_row_key_.empty() && right_row_key_.empty()) {
-      *resource = new BigtableRowRangeResource(cbt::RowRange::InfiniteRange());
-      return Status::OK();
-    }
-
-    // open
-    if (left_open_ && right_open_) {
-      *resource = new BigtableRowRangeResource(
-          cbt::RowRange::Open(left_row_key_, right_row_key_));
-      return Status::OK();
-    }
-    // closed
-    if (!left_open_ && !right_open_) {
-      *resource = new BigtableRowRangeResource(
-          cbt::RowRange::Closed(left_row_key_, right_row_key_));
-      return Status::OK();
-    }
-    // right_open
-    if (!left_open_ && right_open_) {
-      *resource = new BigtableRowRangeResource(
-          cbt::RowRange::RightOpen(left_row_key_, right_row_key_));
-      return Status::OK();
-    }
-    // left_open
-    if (left_open_ && !right_open_) {
-      *resource = new BigtableRowRangeResource(
-          cbt::RowRange::LeftOpen(left_row_key_, right_row_key_));
-      return Status::OK();
-    }
-    return errors::Internal("Reached impossible branch.");
-  }
-
- private:
-  mutable mutex mu_;
-  std::string left_row_key_ TF_GUARDED_BY(mu_);
-  bool left_open_ TF_GUARDED_BY(mu_);
-  std::string right_row_key_ TF_GUARDED_BY(mu_);
-  bool right_open_ TF_GUARDED_BY(mu_);
-};
-
-REGISTER_KERNEL_BUILDER(Name("BigtableRowRange").Device(DEVICE_CPU),
-                        BigtableRowRangeOp);
-
-class BigtablePrintRowRangeOp : public OpKernel {
- public:
-  explicit BigtablePrintRowRangeOp(OpKernelConstruction* context)
-      : OpKernel(context) {}
-
-  void Compute(OpKernelContext* context) override {
-    BigtableRowRangeResource* resource;
-    OP_REQUIRES_OK(context,
-                   GetResourceFromContext(context, "resource", &resource));
-    core::ScopedUnref unref(resource);
-
-    Tensor* output_tensor = NULL;
-    OP_REQUIRES_OK(context, context->allocate_output(0, {1}, &output_tensor));
-    auto output_v = output_tensor->tensor<tstring, 1>();
-
-    output_v(0) = resource->ToString();
-  }
-
- private:
-  mutable mutex mu_;
-};
-
-REGISTER_KERNEL_BUILDER(Name("BigtablePrintRowRange").Device(DEVICE_CPU),
-                        BigtablePrintRowRangeOp);
-
-class BigtableEmptyRowsetOp
-    : public OpKernelCreatingResource<BigtableRowsetResource> {
- public:
-  explicit BigtableEmptyRowsetOp(OpKernelConstruction* ctx)
-      : OpKernelCreatingResource<BigtableRowsetResource>(ctx) {
-    VLOG(1) << "BigtableEmptyRowsetOp ctor ";
-  }
-
- private:
-  Status CreateResource(BigtableRowsetResource** resource)
-      TF_EXCLUSIVE_LOCKS_REQUIRED(mu_) override {
-    *resource = new BigtableRowsetResource(cbt::RowSet());
-    return Status::OK();
-  }
-
- private:
-  mutable mutex mu_;
-};
-
-REGISTER_KERNEL_BUILDER(Name("BigtableEmptyRowset").Device(DEVICE_CPU),
-                        BigtableEmptyRowsetOp);
-
-class BigtablePrintRowsetOp : public OpKernel {
- public:
-  explicit BigtablePrintRowsetOp(OpKernelConstruction* context)
-      : OpKernel(context) {}
-
-  void Compute(OpKernelContext* context) override {
-    BigtableRowsetResource* resource;
-    OP_REQUIRES_OK(context,
-                   GetResourceFromContext(context, "resource", &resource));
-    core::ScopedUnref unref(resource);
-
-    // Create an output tensor
-    Tensor* output_tensor = NULL;
-    OP_REQUIRES_OK(context, context->allocate_output(0, {1}, &output_tensor));
-    auto output_v = output_tensor->tensor<tstring, 1>();
-
-    output_v(0) = resource->ToString();
-  }
-
- private:
-  mutable mutex mu_;
-};
-
-REGISTER_KERNEL_BUILDER(Name("BigtablePrintRowset").Device(DEVICE_CPU),
-                        BigtablePrintRowsetOp);
-
-class BigtableRowsetAppendRowOp : public OpKernel {
- public:
-  explicit BigtableRowsetAppendRowOp(OpKernelConstruction* context)
-      : OpKernel(context) {
-    OP_REQUIRES_OK(context, context->GetAttr("row_key", &row_key_));
-  }
-
-  void Compute(OpKernelContext* context) override {
-    BigtableRowsetResource* resource;
-    OP_REQUIRES_OK(context,
-                   GetResourceFromContext(context, "resource", &resource));
-    core::ScopedUnref unref(resource);
-
-    resource->AppendRow(row_key_);
-  }
-
- private:
-  mutable mutex mu_;
-  std::string row_key_;
-};
-
-REGISTER_KERNEL_BUILDER(Name("BigtableRowsetAppendRow").Device(DEVICE_CPU),
-                        BigtableRowsetAppendRowOp);
-
-class BigtableRowsetAppendRowRangeOp : public OpKernel {
- public:
-  explicit BigtableRowsetAppendRowRangeOp(OpKernelConstruction* context)
-      : OpKernel(context) {}
-
-  void Compute(OpKernelContext* context) override {
-    BigtableRowsetResource* row_set_resource;
-    OP_REQUIRES_OK(context, GetResourceFromContext(context, "row_set_resource",
-                                                   &row_set_resource));
-    core::ScopedUnref row_set_resource_unref(row_set_resource);
-
-    BigtableRowRangeResource* row_range_resource;
-    OP_REQUIRES_OK(context,
-                   GetResourceFromContext(context, "row_range_resource",
-                                          &row_range_resource));
-    core::ScopedUnref row_range_resource_unref(row_range_resource);
-
-    row_set_resource->AppendRowRange(row_range_resource->RowRange());
-  }
-
- private:
-  mutable mutex mu_;
-  std::string row_key_;
-};
-
-REGISTER_KERNEL_BUILDER(Name("BigtableRowsetAppendRowRange").Device(DEVICE_CPU),
-                        BigtableRowsetAppendRowRangeOp);
-
-class BigtablePrefixRowRangeOp
-    : public OpKernelCreatingResource<BigtableRowRangeResource> {
- public:
-  explicit BigtablePrefixRowRangeOp(OpKernelConstruction* ctx)
-      : OpKernelCreatingResource<BigtableRowRangeResource>(ctx) {
-    VLOG(1) << "BigtablePrefixRowRangeOp ctor ";
-    OP_REQUIRES_OK(ctx, ctx->GetAttr("prefix_str", &prefix_str_));
-  }
-
- private:
-  Status CreateResource(BigtableRowRangeResource** resource)
-      TF_EXCLUSIVE_LOCKS_REQUIRED(mu_) override {
-    *resource =
-        new BigtableRowRangeResource(cbt::RowRange::Prefix(prefix_str_));
-    return Status::OK();
-  }
-
- private:
-  mutable mutex mu_;
-  std::string prefix_str_;
-};
-
-REGISTER_KERNEL_BUILDER(Name("BigtablePrefixRowRange").Device(DEVICE_CPU),
-                        BigtablePrefixRowRangeOp);
-
-class BigtableRowsetIntersectOp : public OpKernel {
- public:
-  explicit BigtableRowsetIntersectOp(OpKernelConstruction* context)
-      : OpKernel(context) {}
-
-  void Compute(OpKernelContext* context) override TF_LOCKS_EXCLUDED(mu_) {
-    mutex_lock l(mu_);
-    ResourceMgr* mgr = context->resource_manager();
-    OP_REQUIRES_OK(context, cinfo_.Init(mgr, def()));
-
-    BigtableRowsetResource* row_set_resource;
-    OP_REQUIRES_OK(context, GetResourceFromContext(context, "row_set_resource",
-                                                   &row_set_resource));
-    core::ScopedUnref row_set_resource_unref(row_set_resource);
-
-    BigtableRowRangeResource* row_range_resource;
-    OP_REQUIRES_OK(context,
-                   GetResourceFromContext(context, "row_range_resource",
-                                          &row_range_resource));
-    core::ScopedUnref row_range_resource_unref(row_range_resource);
-
-    BigtableRowsetResource* result_resource;
-    OP_REQUIRES_OK(
-        context,
-        mgr->LookupOrCreate<BigtableRowsetResource>(
-            cinfo_.container(), cinfo_.name(), &result_resource,
-            [this, row_set_resource, row_range_resource](
-                BigtableRowsetResource** ret) TF_EXCLUSIVE_LOCKS_REQUIRED(mu_) {
-              *ret = new BigtableRowsetResource(
-                  row_set_resource->Intersect(row_range_resource->RowRange()));
-              return Status::OK();
-            }));
-
-    OP_REQUIRES_OK(context, MakeResourceHandleToOutput(
-                                context, 0, cinfo_.container(), cinfo_.name(),
-                                TypeIndex::Make<BigtableRowsetResource>()));
-  }
-
- protected:
-  // Variables accessible from subclasses.
-  mutex mu_;
-  ContainerInfo cinfo_ TF_GUARDED_BY(mu_);
-};
-
-REGISTER_KERNEL_BUILDER(Name("BigtableRowsetIntersect").Device(DEVICE_CPU),
-                        BigtableRowsetIntersectOp);
 
 }  // namespace
 }  // namespace data
